fix : folder페이지 url추가 위치 오류 수정
